Rename component name
